restructure p4fmt dir
